pyproject.toml
[tool.poetry]
name = "cloudracehub"
version = "0.0.1"
description = "cloud race hub"
authors = ["WALLE-AI <gaojing850063636@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.55.3"
transformers = "^4.46.3"
streamlit = "^1.40.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
pdfplumber = "^0.11.4"
tiktoken = "^0.8.0"
pydantic-extra-types = "^2.10.0"
gradio = "4.44.0"
oss2 = "^2.19.1"
dashscope = "^1.20.13"
qwen-agent = {extras = ["rag"], version = "^0.0.11"}
pyaudio = "^0.2.14"
opencv-python = "^4.10.0.84"
peft = "^0.13.2"
swanlab = "^0.3.27"
datasets = "^3.1.0"
json-repair = "^0.30.2"
modelscope-studio = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
